refact modeInterface
