+++ src/main.ts
@@ -4,17 +4,15 @@ import { onDocumentMouseMove, onDocumentMouseUp, onDocumentClick, onDocumentMous
 import { Mode } from "./modeInterface";
 import { SelectMode } from "./selectMode";
 import { textMode } from "./textMode";
-import { elementVoid, elementOpen, elementClose, patch } from "incremental-dom";
+import { patch } from "incremental-dom";
 import { MenuListComponent, ModeName } from "./menuComponent";
 import { Component, WindowComponent } from "./component";
 import { SvgContainerComponent } from "./svgContainerComponent";
 import { StyleConfigComponent } from "./styleConfigComponent";
 import { el } from "./utils";
 import { collectPaintServer } from "./paintServer";
-import { NodeMode } from "./nodeMode";
-import { multiShaper, shaper } from "./shapes";
+import { shaper } from "./shapes";
 import { LoadedImage, collectImages } from "./imageHelpters";
-import uuidStatic from "uuid";
 
 declare function acquireVsCodeApi(): any;
 
@@ -57,13 +55,13 @@ class ContentChildrenComponent implements Component {
 
     render() {
         el`header`;
-        this.menuListComponent.render();
+            this.menuListComponent.render();
         el`/header`;
         el`div :key="body" *class="svgeditor-body"`;
-        this.svgContainerComponent.render();
+            this.svgContainerComponent.render();
         el`/div`;
         el`footer`;
-        this.styleConfigComponent.render();
+            this.styleConfigComponent.render();
         el`/footer`;
     }
 }
