+++ src/rectMode.ts
@@ -21,7 +21,7 @@ export class RectMode extends Mode {
             const root = svgVirtualMap[uu];
             event.stopPropagation();
             this.isDragging = true;
-            this.startCursorPos = vfp(this.inTargetCoordinate(this.cursor(event), uu));
+            this.startCursorPos = vfp(this.inTargetCoordinate(this.cursor(event), [uu]));
             this.dragTargetUuid = uuidStatic.v4();
             if (root.tag === "svg") {
                 const pe: ParsedElement = {
@@ -52,7 +52,7 @@ export class RectMode extends Mode {
     }
     onDocumentMouseMove(event: MouseEvent): void {
         if (this.isDragging && this.startCursorPos && this.dragTargetUuid) {
-            const {x: cx, y: cy} = this.inTargetCoordinate(this.cursor(event), this.dragTargetUuid)
+            const {x: cx, y: cy} = this.inTargetCoordinate(this.cursor(event), [this.dragTargetUuid])
             const leftTop = v(Math.min(cx, this.startCursorPos.x), Math.min(cy, this.startCursorPos.y));
             const size = v(Math.abs(cx - this.startCursorPos.x), Math.abs(cy - this.startCursorPos.y));
             shaper(this.dragTargetUuid).size = size;
@@ -65,12 +65,5 @@ export class RectMode extends Mode {
     }
     onDocumentMouseLeave(event: Event): void {
         this.onDocumentMouseUp();
-    }
-
-    /**
-     * Transform a (mouse) point into that in coordinate of a target shape by inverse mapping.
-     */
-    private inTargetCoordinate(point: Point, targetUuid: string): Point {
-        return applyToPoint(inverse(shaper(targetUuid).allTransform()), point);
     }
 }
