+++ src/selectMode.ts
@@ -109,126 +109,6 @@ export class SelectMode extends Mode {
     onDocumentMouseLeave(event: Event) {
         this.onDocumentMouseUp();
     }
-    onOperatorClicked(name: OperatorName) {
-        super.onOperatorClicked(name);
-        const uuids = this._selectedShapeUuids;
-        switch (name) {
-            case "duplicate":
-            if (uuids && this.commonParent) {
-                let parentPe = svgVirtualMap[this.commonParent];
-                const copiedUuids: string[] = [];
-                for (let uuid of uuids) {
-                    const pe = svgVirtualMap[uuid];
-                    const copied = deepCopy(pe);
-                    copied.uuid = uuidStatic.v4();
-                    copiedUuids.push(copied.uuid);
-                    if ("children" in parentPe) {
-                        parentPe.children.push(copied);
-                    }
-                }
-                refleshContent();       // make real elements
-                let tmp: null | OneOrMore<string> = null;
-                for (let copiedUuid of copiedUuids) {
-                    const fourPercentX = convertToPixel({unit: "%", value: 4, attrName: "x"}, copiedUuid);
-                    const fourPercentY = convertToPixel({unit: "%", value: 4, attrName: "y"}, copiedUuid);
-                    shaper(copiedUuid).move(v(fourPercentX, fourPercentY));
-                    tmp ? tmp.push(copiedUuid) : tmp = [copiedUuid];
-                }
-                this.selectedShapeUuids = tmp;
-            }
-            break;
-            case "delete":
-            if (uuids && this.commonParent) {
-                let parentPe = svgVirtualMap[this.commonParent];
-                if ("children" in parentPe) {
-                    parentPe.children = parentPe.children.filter(c => uuids.indexOf(c.uuid) === -1);
-                    this.selectedShapeUuids = null;
-                }
-            }
-            break;
-            case "group":
-            if (uuids && this.commonParent) {
-                let parentPe = svgVirtualMap[this.commonParent];
-                if ("children" in parentPe) {
-                    const childrenPe = uuids.map(uu => svgVirtualMap[uu]);
-                    const groupUuid = uuidStatic.v4();
-                    childrenPe.forEach(c => c.parent = groupUuid);
-                    parentPe.children.push({
-                        uuid: groupUuid,
-                        tag: "g",
-                        isRoot: false,
-                        parent: this.commonParent,
-                        attrs: {
-                            ...Mode.baseAttrsDefaultImpl(),
-                            ...Mode.presentationAttrsDefaultImpl(),
-                            fill: null,
-                            stroke: null
-                        },
-                        children: childrenPe
-                    });
-                    parentPe.children = parentPe.children.filter(c => uuids.indexOf(c.uuid) === -1);
-                    refleshContent();       // make real elements
-                    this.selectedShapeUuids = [groupUuid];
-                }
-            }
-            break;
-            case "ungroup":
-            if (uuids && uuids.length === 1) {
-                const pe = svgVirtualMap[uuids[0]];
-                if (pe.tag === "g" && pe.parent) {
-                    const gParent = pe.parent;
-                    const parentPe = svgVirtualMap[gParent];
-                    for (let c of pe.children) {
-                        c.parent = gParent;
-                        if ("children" in parentPe) parentPe.children.push(c);
-                        if (pe.attrs.transform) shaper(c.uuid).appendTransformDescriptors(pe.attrs.transform.descriptors, "left");
-                    }
-                    if ("children" in parentPe) parentPe.children = parentPe.children.filter(c => c.uuid !== pe.uuid);
-                    this.selectedShapeUuids = null;
-                }
-            }
-            break;
-            case "bring forward":
-            if (uuids) {
-                traverse(svgdata, (pe, parentPe, index) => {
-                    let prePe: ParsedElement;
-                    if (index !== null && index >= 1 && parentPe && (prePe = parentPe.children[index - 1]) && uuids.indexOf(pe.uuid) === -1) {
-                        parentPe.children[index - 1] = pe;
-                        parentPe.children[index] = prePe;
-                    }
-                });
-            }
-            break;
-            case "send backward":
-            if (uuids) {
-                traverse(svgdata, (pe, parentPe, index) => {
-                    let prePe: ParsedElement;
-                    if (index !== null && index >= 1 && parentPe && (prePe = parentPe.children[index - 1]) && uuids.indexOf(pe.uuid) !== -1) {
-                        parentPe.children[index - 1] = pe;
-                        parentPe.children[index] = prePe;
-                    }
-                });
-            }
-            break;
-            case "align left":
-            this.align("left");
-            break;
-            case "align right":
-            this.align("right");
-            break;
-            case "align top":
-            this.align("top");
-            break;
-            case "align bottom":
-            this.align("bottom");
-            break;
-            case "object to path":
-            if (uuids) multiShaper(uuids).toPath();
-            break;
-        }
-        refleshContent();
-        sendBackToEditor();
-    }
 
     render() {
         // Decorate groups.
@@ -306,64 +186,6 @@ export class SelectMode extends Mode {
             this.startCursorPos = vfp(this.inTargetCoordinate(this.cursor(event), this._selectedShapeUuids));
             this.startShapeSize = multiShaper(this._selectedShapeUuids).size;
             this.previousRawCursorPos = this.cursor(event);
-        }
-    }
-
-    /**
-     * Transform a (mouse) point into that in coordinate of a target shape by inverse mapping.
-     */
-    private inTargetCoordinate(point: Point, targetUuids: OneOrMore<string>): Point {
-        return applyToPoint(inverse(multiShaper(targetUuids).allTransform()), point);
-    }
-
-    private escapeToNormalCoordinate(point: Point, targetUuids: OneOrMore<string>): Point {
-        return applyToPoint(multiShaper(targetUuids).allTransform(), point);
-    }
-
-    private align(dir: "left" | "right" | "top" | "bottom"): void {
-        const minOrMax = (dir === "left" || dir === "top") ? Math.min : Math.max;
-        const xOrY = (dir === "left" || dir === "right") ? "x" : "y";
-        type Corner = "leftTop" | "leftBottom" | "rightTop" | "rightBottom";
-        const cornerNames =
-            (<() => [Corner, Corner]>(() => {
-                switch (dir) {
-                    case "left":
-                    return ["leftTop", "leftBottom"];
-                    case "right":
-                    return ["rightTop", "rightBottom"];
-                    case "top":
-                    return ["leftTop", "rightTop"];
-                    case "bottom":
-                    return ["leftBottom", "rightBottom"];
-                }
-            }))();
-        if (this._selectedShapeUuids) {
-            let edge: number | null = null;
-            const corners: {[uuid: string]: {cornerName: Corner, escapedPoint: Point}} = {};
-            for (let uuid of this._selectedShapeUuids) {
-                const [nameFirst, nameSecond] = cornerNames;
-                let escapedCornerFirst = this.escapeToNormalCoordinate(shaper(uuid)[nameFirst], [uuid]);
-                let escapedCornerSecond = this.escapeToNormalCoordinate(shaper(uuid)[nameSecond], [uuid]);
-                let tmp: number;
-                if ((tmp = minOrMax(escapedCornerFirst[xOrY], escapedCornerSecond[xOrY])) === escapedCornerFirst[xOrY]) {
-                    corners[uuid] = {
-                        cornerName: nameFirst,
-                        escapedPoint: escapedCornerFirst
-                    }
-                } else {
-                    corners[uuid] = {
-                        cornerName: nameSecond,
-                        escapedPoint: escapedCornerSecond
-                    }
-                }
-                edge = edge !== null && minOrMax(tmp, edge) || tmp;
-            }
-            if (edge !== null) for (let uuid of this._selectedShapeUuids) {
-                const cornersData = corners[uuid];
-                cornersData.escapedPoint[xOrY] = edge;
-                const targetCorner = this.inTargetCoordinate(cornersData.escapedPoint, [uuid]);
-                shaper(uuid)[cornersData.cornerName] = vfp(targetCorner);
-            }
         }
     }
 }
